Move SendErrorFilter in error package
